+++ client/src/components/ClassForm/FormSections/ImageUpload.scss
